[Fix]: exclude functions and objects that were using vitest from tsconfig for docker build
